requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r requirements.in
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.6
    # via pytest-watch
coverage[toml]==7.2.3
    # via pytest-cov
django==4.2
    # via
    #   -r requirements.in
    #   django-extensions
    #   djangorestframework
django-extensions==3.2.3
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
docopt==0.6.2
    # via pytest-watch
exceptiongroup==1.1.1
    # via pytest
flake8==6.0.0
    # via -r requirements.in
importlib-metadata==4.13.0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
kombu==5.2.4
    # via celery
mccabe==0.7.0
    # via flake8
packaging==23.1
    # via pytest
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-watch
pytest-cov==4.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-watch==4.2.0
    # via -r requirements.in
pytz==2023.3
    # via
    #   celery
    #   djangorestframework
redis==4.5.4
    # via -r requirements.in
six==1.16.0
    # via click-repl
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   coverage
    #   pytest
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchdog==3.0.0
    # via pytest-watch
wcwidth==0.2.6
    # via prompt-toolkit
zipp==3.15.0
    # via importlib-metadata
